+++ app/src/main/java/com/pzbdownloaders/add_bullet_points_note_feature/presentation/components/SingleRowBulletPoint.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material.CheckboxDefaults
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.icons.Icons
@@ -86,7 +87,11 @@ fun SingleRowBulletPoint(
                 unfocusedTextColor = MaterialTheme.colors.onPrimary,
                 unfocusedIndicatorColor = MaterialTheme.colors.primary,
                 focusedIndicatorColor = MaterialTheme.colors.primary,
-                cursorColor = MaterialTheme.colors.onPrimary
+                cursorColor = MaterialTheme.colors.onPrimary,
+                selectionColors = TextSelectionColors(
+                    handleColor = MaterialTheme.colors.onPrimary,
+                    backgroundColor = Color.Gray
+                )
             ),
             textStyle = LocalTextStyle.current.copy(
                 fontFamily = FontFamily.fontFamilyRegular