+++ app/src/main/java/com/pzbdownloaders/scribble/add_checkbox_note_feature/presentation/components/SingleRowCheckBox.kt
@@ -14,6 +14,7 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material.CheckboxDefaults
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.icons.Icons
@@ -110,7 +111,11 @@ fun SingleRowCheckBox(
                 unfocusedTextColor = MaterialTheme.colors.onPrimary,
                 unfocusedIndicatorColor = MaterialTheme.colors.primary,
                 focusedIndicatorColor = MaterialTheme.colors.primary,
-                cursorColor = MaterialTheme.colors.onPrimary
+                cursorColor = MaterialTheme.colors.onPrimary,
+                selectionColors = TextSelectionColors(
+                    handleColor = MaterialTheme.colors.onPrimary,
+                    backgroundColor = Color.Gray
+                )
             ),
             textStyle = LocalTextStyle.current.copy(
                 fontFamily = FontFamily.fontFamilyRegular