+++ app/src/main/java/com/pzbdownloaders/scribble/edit_note_feature/presentation/components/MainStructureEditNote.kt
@@ -741,7 +741,8 @@ fun MainStructureEditNote(
                         isUnderlineActivated = isUnderlineActivated,
                         isItalicActivated = isItalicActivated,
                         isOrderedListActivated = isOrderedListActivated,
-                        isUnOrderedListActivated = isUnOrderedListActivated
+                        isUnOrderedListActivated = isUnOrderedListActivated,
+                        mutableStateOf(false)
                     )
                 }
             }