+++ app/src/main/java/com/pzbdownloaders/scribble/locked_notes_feature/presentation/components/MainStructureAddNoteLockedScreen.kt
@@ -84,7 +84,6 @@ import com.pzbdownloaders.scribble.common.presentation.MainActivityViewModel
 import com.pzbdownloaders.scribble.common.presentation.Screens
 import com.pzbdownloaders.scribble.common.presentation.components.AlertDialogBoxTrialEnded
 import com.pzbdownloaders.scribble.main_screen.domain.model.Note
-import io.ktor.util.reflect.typeInfoImpl
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch