+++ app/src/main/java/com/pzbdownloaders/scribble/main_screen/presentation/components/CameraScreen.kt
@@ -1,0 +1,61 @@
+package com.pzbdownloaders.scribble.main_screen.presentation.components
+
+import android.graphics.Color
+import android.widget.LinearLayout
+import android.widget.ListPopupWindow.MATCH_PARENT
+import androidx.camera.core.ImageCapture
+import androidx.camera.core.ImageProxy
+import androidx.camera.view.LifecycleCameraController
+import androidx.camera.view.PreviewView
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExtendedFloatingActionButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.compose.LocalLifecycleOwner
+import androidx.navigation.NavHostController
+
+@Composable
+fun CameraScreen(navHostController: NavHostController) {
+
+    val context = LocalContext.current
+    val lifeCycleOwner = LocalLifecycleOwner.current
+    val cameraController = remember { LifecycleCameraController(context) }
+    Scaffold(floatingActionButton = {
+        ExtendedFloatingActionButton(onClick = {
+            cameraController.takePicture(ContextCompat.getMainExecutor(context),
+                object : ImageCapture.OnImageCapturedCallback() {
+                    override fun onCaptureSuccess(image: ImageProxy) {
+                        super.onCaptureSuccess(image)
+                        image.toBitmap()
+                    }
+                })
+        }) {
+                Text("Take Photo")
+        }
+    }) { paddvingValues ->
+        AndroidView(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddvingValues),
+            factory = { context ->
+
+                PreviewView(context).apply {
+                    layoutParams = LinearLayout.LayoutParams(MATCH_PARENT, MATCH_PARENT)
+                    setBackgroundColor(Color.BLACK)
+                    scaleType = PreviewView.ScaleType.FILL_START
+                    implementationMode = PreviewView.ImplementationMode.COMPATIBLE
+                }.also { previewView ->
+                    previewView.controller = cameraController
+                    cameraController.bindToLifecycle(lifeCycleOwner)
+
+                }
+            })
+    }
+}