+++ app/src/main/java/com/pzbdownloaders/scribble/settings_feature/screen/presentation/components/EmailWillBeSendAlertBox.kt
@@ -36,7 +36,6 @@ import com.pzbdownloaders.scribble.common.presentation.FontFamily
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import okhttp3.Dispatcher
 
 @Composable
 fun EmailWillBeSendAlertBox(